# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Café Run Club Rewards API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /runs:
    get:
      tags: [runs]
      operationId: listMyRuns
      summary: List my runs
      responses:
        '200':
          description: List of runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Run'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [runs]
      operationId: logRun
      summary: Log a new run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunInput'
      responses:
        '201':
          description: Run logged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /runs/{runId}:
    delete:
      tags: [runs]
      operationId: deleteRun
      summary: Delete a run
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Run deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteSuccess'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DeleteSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    Run:
      type: object
      required: [id, userId, distanceMiles, date, createdAt, source]
      properties:
        id:
          type: integer
        userId:
          type: string
        distanceMiles:
          type: number
        date:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        source:
          type: string
          enum: [manual, strava]
          default: manual
        stravaActivityId:
          type: ["string", "null"]
        clubEventId:
          type: ["integer", "null"]
...
    RunInput:
      type: object
      required: [distanceMiles, date]
      properties:
        distanceMiles:
          type: number
          minimum: 0.01
        date:
          type: string
          format: date
        notes:
          type: string
        forUserId:
          type: string
          description: Optional userId of a dependent to log the run for
...